chore: rename marathon/savebar to marathon/stepperBar for clear meaning
